extract TrackedEntitySearchPagingItem to its own file
